+++ views/js/front.js
@@ -1,5 +1,5 @@
 /**
- * 2007-2019 Ingenico
+ * 2007-2021 Ingenico
  *
  * NOTICE OF LICENSE
  *
@@ -12,7 +12,7 @@
  * to contact@ingenico.com we can send you a copy immediately.
  *
  *  @author    Ingenico <contact@ingenico.com>
- *  @copyright 2007-2019 Ingenico
+ *  @copyright 2007-2021 Ingenico
  *  @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
  */
 "use strict"
@@ -85,6 +85,114 @@ const Ingenico = {
     },
 };
 
+window.IngenicoCCInline = {
+    iframe: null,
+    wrap: null,
+    oc_wrap: null,
+    tos: null,
+
+    init: function() {
+        var self = this;
+
+        this.iframe = $('iframe#ingenico_iframe');
+        this.wrap = $('.iframe-wrap');
+        this.oc_wrap = $('.one-click-wrap');
+        this.tos = $('#conditions-to-approve input[type="checkbox"]');
+
+        // Hide iframe if ToS has been unchecked
+        if (this.tos.length > 0) {
+            this.watchTos();
+        }
+
+        // Hide "payment confirmation" on the checkout page when cc form chosen
+        $(document).on('click', '.payment-options [name="payment-option"]', function() {
+            // If Inline CC mode is active
+            if ($('#pay-with-' + $(this).prop('id') + '-form').find(self.iframe).length > 0) {
+                setTimeout(function () {
+                    self.hideConfirmationButton();
+                }, 1500);
+            } else {
+                self.showConfirmationButton();
+            }
+        });
+
+        // Toggle CCForm on alias selection
+        $(document).on('click', '.payment-options input[name="alias"]', function() {
+            // Check if the cc iframe is exists
+            if (self.iframe.length === 0) {
+                return;
+            }
+
+            if ($(this).val() === 'new') {
+                self.showOneClickIframe();
+                self.hideConfirmationButton();
+            } else {
+                self.hideOneClickIframe();
+                self.showConfirmationButton();
+            }
+        });
+    },
+
+    /**
+     * Watch "Terms of service" and shows or hides iframe depends on the checkbox
+     */
+    watchTos: function () {
+        var self = this;
+
+        if (self.tos.length > 0) {
+            setInterval(function () {
+                if (self.tos.is(':checked')) {
+                    self.showIframe();
+                } else {
+                    self.hideIframe();
+                }
+            }, 500);
+        }
+    },
+
+    /**
+     * Hide CC Iframe
+     */
+    hideIframe: function () {
+        this.wrap.hide();
+    },
+
+    /**
+     * Show CC Iframe
+     */
+    showIframe: function () {
+        this.wrap.show();
+    },
+
+    /**
+     * Hide "One Click" iframe
+     */
+    hideOneClickIframe: function () {
+        this.oc_wrap.hide();
+    },
+
+    /**
+     * Show "One Click" iframe
+     */
+    showOneClickIframe: function () {
+        this.oc_wrap.show();
+    },
+
+    /**
+     * Hide Confirmation Button
+     */
+    hideConfirmationButton: function () {
+        $('#payment-confirmation button').hide();
+    },
+
+    /**
+     * Show Confirmation Button
+     */
+    showConfirmationButton: function () {
+        $('#payment-confirmation button').show();
+    }
+};
+
 $(function() {
     Ingenico.init();
     Ingenico.setEnvironment();
@@ -94,29 +202,4 @@ jQuery(document).ready(function ($) {
     $(document).on('ingenico:inline:failure', 'body', function(e, msg, aliasId, cardBrand, url) {
         Ingenico.inlineFailure(e, msg, aliasId, cardBrand, url);
     });
-
-    // Hide "payment confirmation" on the checkout page when cc form chosen
-    $(document).on('click', '.payment-options [name="payment-option"]', function() {
-        if ($('#pay-with-' + $(this).prop('id') + '-form').find('iframe#ingenico_iframe').length > 0) {
-            setTimeout(function () {
-                $('#payment-confirmation').hide();
-            }, 3000);
-        } else {
-            $('#payment-confirmation').show();
-        }
-    });
-
-    $(document).on('click', '.payment-options [name="alias"]', function() {
-        // Check if the cc iframe is exists
-        if ($('iframe#ingenico_iframe').length === 0) {
-            return;
-        }
-
-        if ($(this).val() === 'new') {
-            $('#payment-confirmation').hide();
-        } else {
-            $('#payment-confirmation').show();
-        }
-    });
-
 });